Emax_code
The main bcrm function with the 4-parameter Emax model incorporated as one of the choices of the model with prior parameters specified as in the paper and the randomization to the control arm. The file also includes the code for computing the benchmark estimates. This file should be run before the file ""emax performance example"