BCRM package with the control group
The includes the original bcrm function with the extension allowing randomization to the control group.  The randomization ratio can be customized using parameters "cohort" (for the investigational arm) and "cohort.control" (for the control arm). The rest of the functions are as in the original package. This file should be run before "L2R example"